renamed Main.js to Home.js
cleaned up code, created a separate component for card details
